requirements.txt
annotated-types==0.5.0
Appium-Python-Client==3.0.0
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
exceptiongroup==1.1.3
h11==0.14.0
idna==3.4
outcome==1.2.0
pyaes==1.6.1
pyasn1==0.5.0
pydantic==2.3.0
pydantic_core==2.6.3
PySocks==1.7.1
python-dotenv==1.0.0
requests==2.31.0
rsa==4.9
selenium==4.12.0
smsactivate==1.5
sniffio==1.3.0
sortedcontainers==2.4.0
Telethon==1.30.3
trio==0.22.2
trio-websocket==0.10.4
typing_extensions==4.7.1
urllib3==2.0.4
wsproto==1.2.0
